Initial rev
A tool to fix the path to python in scripts made by virtualenv. It'll help if you have spaces in the file path, in which case the scripts built by virtualenv break. This replaces the invocation of python directly with an invocation of /usr/bin/env.exe which then invokes python using the activated environment's path.